Partition to modules
